requirements.txt
cpbd
dash
face_recognition
huggingface_hub
matplotlib
mtcnn
numpy
opencv_python
opencv_python_headless
pandas
piexif
Pillow
plotly
psutil
PyQt6
PyQt6_sip
scikit-image
tiktoken
tqdm
tweepy
brotli
tweepy
requests>=2.31.0
tomli>=2.0.1; python_version<"3.11"
tomli-w>=1.0.0; python_version<"3.11"